fix: rearrange separate functions and types into one  service file
