+++ src/services/AuthService/types.ts
@@ -1,0 +1,12 @@
+export interface TokenResponse {
+  access_token: string;
+  expires_in: number;
+  scope: string;
+  token_type: string;
+  refresh_token?: string;
+}
+
+export interface FormProps {
+  email: string;
+  password: string;
+}
